app-audio/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic<3
torch==2.1.0+cu121
openai-whisper>=20231117
numpy>=1.24.0
ffmpeg-python>=0.2.0
